requirements.txt
faiss=1.7.0
numpy=1.19.5
pandas=1.3.2
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-cluster=1.5.9=py38_torch_1.9.0_cu102
pytorch-scatter=2.0.8=py38_torch_1.9.0_cu102
pytorch-sparse=0.6.12=py38_torch_1.9.0_cu102
pytorch-spline-conv=1.2.1=py38_torch_1.9.0_cu102
scikit-learn=0.24.2